refactor: ViewHolder 클래스를 파일로 따로 관리
