+++ app/src/main/java/woowacourse/movie/MovieListAdapter.kt
@@ -5,9 +5,6 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.BaseAdapter
-import android.widget.Button
-import android.widget.ImageView
-import android.widget.TextView
 
 class MovieListAdapter(private val movies: List<Movie>) : BaseAdapter() {
     override fun getCount(): Int = movies.size
@@ -18,16 +15,16 @@ class MovieListAdapter(private val movies: List<Movie>) : BaseAdapter() {
 
     override fun getView(position: Int, convertView: View?, parent: ViewGroup?): View {
         val view: View
-        val viewHolder: ViewHolder
+        val viewHolder: MovieListViewHolder
 
         if (convertView == null) {
             view = LayoutInflater.from(parent?.context)
                 .inflate(R.layout.item_movie_list, parent, false)
-            viewHolder = ViewHolder()
+            viewHolder = MovieListViewHolder()
             viewHolder.initView(view)
             view.tag = viewHolder
         } else {
-            viewHolder = convertView.tag as ViewHolder
+            viewHolder = convertView.tag as MovieListViewHolder
             view = convertView
         }
         viewHolder.setData(getItem(position))
@@ -41,28 +38,5 @@ class MovieListAdapter(private val movies: List<Movie>) : BaseAdapter() {
 
     companion object {
         const val MOVIE_DATA_KEY = "movieData"
-    }
-
-    private class ViewHolder {
-        var poster: ImageView? = null
-        var title: TextView? = null
-        var releaseDate: TextView? = null
-        var runningTime: TextView? = null
-        var bookButton: Button? = null
-
-        fun initView(view: View) {
-            poster = view.findViewById(R.id.iv_movie_poster)
-            title = view.findViewById(R.id.tv_movie_title)
-            releaseDate = view.findViewById(R.id.tv_movie_release_date)
-            runningTime = view.findViewById(R.id.tv_movie_running_time)
-            bookButton = view.findViewById(R.id.bt_book_now)
-        }
-
-        fun setData(movieData: Movie) {
-            poster?.setImageResource(movieData.poster)
-            title?.text = movieData.title
-            releaseDate?.text = movieData.releaseDate
-            runningTime?.text = movieData.runningTime
-        }
     }
 }
